JOTS-35-Course-Management-Database-Side

+++ jots-frontend/src/app/users/list.component.ts
@@ -1,13 +1,14 @@
 ﻿import { Component, OnInit } from '@angular/core';
-import { first } from 'rxjs/operators';
+import { first, map } from 'rxjs/operators';
 
 import { AccountService } from '@app/_services';
+import { User } from '@app/_models';
 
 @Component({ templateUrl: 'list.component.html' })
 export class ListComponent implements OnInit {
     users?: any[];
 
-    constructor(private accountService: AccountService) {}
+    constructor(private accountService: AccountService) { }
 
     ngOnInit() {
         this.accountService.getAll()
@@ -16,10 +17,18 @@ export class ListComponent implements OnInit {
     }
 
     deleteUser(id: string) {
-        const user = this.users!.find(x => x.id === id);
-        user.isDeleting = true;
-        this.accountService.delete(id)
-            .pipe(first())
-            .subscribe(() => this.users = this.users!.filter(x => x.id !== id));
+        const user = this.users?.find(x => x.id === id);
+
+        if (user) {
+            user.isDeleting = true;
+
+            this.accountService.delete(id)
+                .pipe(first())
+                .subscribe(() => {
+                    this.users = this.users?.filter(x => x.id !== id);
+                    delete user.isDeleting; // Remove the isDeleting property
+                });
+        }
     }
+
 }
